norminette almost done
