+++ execution/build_tree.c
@@ -6,25 +6,26 @@
 /*   By: yloutfi <yloutfi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/24 11:25:16 by yloutfi           #+#    #+#             */
-/*   Updated: 2023/07/09 16:41:36 by yloutfi          ###   ########.fr       */
+/*   Updated: 2023/07/10 16:37:38 by yloutfi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minishell.h"
 
-t_exec	*init_exec()
+t_exec	*init_exec(void)
 {
 	t_exec	*exec_node;
-	
+
 	exec_node = malloc(sizeof(t_exec));
 	ft_memset(exec_node, 0, sizeof(t_exec));
 	exec_node->mask = EXEC_NODE;
 	return (exec_node);
 }
+
 t_mask	*build_pipe(t_mask *left, t_mask *right)
 {
 	t_pipe		*pipe_node;
-	
+
 	pipe_node = malloc(sizeof(t_pipe));
 	ft_memset(pipe_node, 0, sizeof(t_pipe));
 	pipe_node->mask = PIPE_NODE;
@@ -32,17 +33,18 @@ t_mask	*build_pipe(t_mask *left, t_mask *right)
 	pipe_node->right = right;
 	return ((t_mask *)pipe_node);
 }
+
 t_mask	*build_exec(t_list *_buffer)
 {
 	t_exec		*exec_node;
 	t_buffer	*buff_node;
 
 	exec_node = init_exec();
-	while(_buffer)
+	while (_buffer)
 	{
 		buff_node = (t_buffer *)_buffer->content;
 		if (buff_node->type == 7)
-			break;
+			break ;
 		else if (buff_node->type == 1)
 			exec_node->cmd = ft_strdup(buff_node->str);
 		else if (buff_node->type == 2)
@@ -60,9 +62,9 @@ t_mask	*build_exec(t_list *_buffer)
 
 t_list	*pipe_finder(t_list *_buffer)
 {
-	t_list	*tmp;
+	t_list		*tmp;
 	t_buffer	*buff;
-	
+
 	tmp = _buffer;
 	while (tmp)
 	{
